pyproject.toml
[tool.poetry]
name = "floral-oracle"
version = "0.1.0"
description = ""
authors = ["Sarosh Manzoor <sarosh.manzoor@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
torch = "^2.0.0"
llama-index = "^0.10.42"
llama-index-embeddings-huggingface = "^0.2.1"
llama-index-llms-llama-cpp = "^0.1.3"
langchain-community = "^0.2.1"
pymupdf = "^1.24.5"

#llama-2-7b-chat = {"url"= "https://huggingface.co/TheBloke/Llama-2-7B-Chat-GGUF/resolve/main/llama-2-7b-chat.Q4_0.gguf"}
langchain-text-splitters = "^0.2.1"
faiss-cpu = "^1.8.0"
langchainhub = "^0.1.17"
panel = "^1.4.4"

[tool.poetry.group.dev.dependencies]
black = { extras = ["d"], version = "^24.4.2" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
